Move log operation start into util
