+++ pyinfra/api/util.py
@@ -209,6 +209,32 @@ def print_host_combined_output(host, combined_output_lines):
             )
 
 
+def log_operation_start(op_meta, op_types=None, prefix="--> "):
+    op_types = op_types or []
+    if op_meta["serial"]:
+        op_types.append("serial")
+    if op_meta["run_once"]:
+        op_types.append("run once")
+
+    args = ""
+    if op_meta["args"]:
+        args = "({0})".format(", ".join(str(arg) for arg in op_meta["args"]))
+
+    logger.info(
+        "{0} {1} {2}".format(
+            click.style(
+                "{0}Starting{1}operation:".format(
+                    prefix,
+                    " {0} ".format(", ".join(op_types)) if op_types else " ",
+                ),
+                "blue",
+            ),
+            click.style(", ".join(op_meta["names"]), bold=True),
+            args,
+        ),
+    )
+
+
 def log_error_or_warning(host, ignore_errors, description="", continue_on_error=False):
     log_func = logger.error
     log_color = "red"
